Fixed AMD64 compilation
git-svn-id: svn://falconpl.org/falcon/core/trunk@555 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/src_lexer.cpp
@@ -440,7 +440,7 @@ int SrcLexer::lex_normal()
          else {
             delete m_in; // all the streams except first are to be disposed.
             m_in = (Stream *) m_streams.back();
-            m_line = (uint32) m_streamLines.back();
+            m_line = (uint32)(int64) m_streamLines.back();
             m_streamLines.popBack();
             m_previousLine = m_line-1;
             next_loop = true;
